Issue #2: Add support for cloning class instances that inherit from Map or Set.
This is achieved by:

  1. Cloning the instance using existing functionality.
  2. Copying the prototype from the original object.
  3. Copying all (own) properties from the source to the target.
  4. Cloning the instance as if it were an object.

This also required splitting most of "cloneObject()" into an extra function, "cloneObjectToTarget", so we can pass in an existing target (i.e. cloned) object. (Ideally, we'd move the value dispatch logic into a separate function, which could help with Issue #1, where you can't pass a Map in as the root object to clone.)

TODO:
- Consider not copying a property if it already exists on the target.
- Re-run benchmarks, as this extra cloning will probably slow things down when cloning Sets/Maps. (As an optimisation, maybe we can skip the extra cloning steps if the original object's prototype is just "Map" or "Set"?)