Add TypeScript declarations.
This allows importing using the syntax "import deepClone = require('fast-deepclone')". It does not support the syntax "import * as deepClone from 'fast-deepclone'". This could be achieved by adding a dummy namespace with the same name as the declared function, e.g. "namespace fastDeepClone {}", but I noticed this doesn't play nicely with an IDE like IntelliJ IDEA - it wasn't picking up parameter hints, and maybe the JSDoc.